# swagger.yaml
openapi: 3.0.0
info:
  title: Submissions API
  contact: {}
  version: '1.0'
servers:
  - url: https://acolyte-submissions-api.herokuapp.com/api/v1
    variables: {}
paths:
...
  /submissions:
    post:
      tags:
        - Submissions
      summary: Create Submitted
      operationId: CreateSubmitted
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
            example:
              company: google
              position: front-end developer
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - Submissions
      summary: Get All Submissions
      operationId: GetAllSubmissions
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /submissions/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The submission ID
    get:
      tags:
        - Submissions
      summary: Get Single Submission
      operationId: GetSingleSubmission
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false

    patch:
      tags:
        - Submissions
      summary: Update Submission
      operationId: UpdateSubmission
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubmissionRequest'
            example:
              company: random
              position: intern
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
        - Submissions
      summary: Delete Submission
      operationId: DeleteSubmission
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
components:
  schemas:
...
    CreateSubmissionRequest:
      title: CreateSubmissionRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: google
        position: front-end developer
    UpdateSubmissionRequest:
      title: UpdateSubmissionRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: random
        position: intern
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...